Serializer v0.1
Cleaned things up around the serializer functionality and just general cleanup of some of the GameObject child classes. Restructured them a bit, added better notes about the functions, and what not. Now, the game object classes (minus Player) all have a new constructor that takes a keyword, uses that keyword to find a json entry in a json, then makes a new object based on that found entry. Another feature to make it easier to create objects as json entries, is that when referencing another object, say a Mod child on an Item, you only need to reference it's keyword. It will then use the keyword constructor to make that child object.

What this does not have is any sort of exception handling, which is going to be required if most of the objects are going to be made in json entries so they can only be validated during runtime when the file is accessed.