serealizer v1.0
Now that I've implemented the try/catch exceptions, I'm OK with calling this feature complete for now. I may eventually change the catch logic for these so they don't completely terminated the program. Maybe they could possible do an auto save or something? I'm not sure. But for now, it will throw an exception and terminate the program with a exit code of 1.
There is one thing to be mindful of when messing with any of the serialization stuff, and that's JSON files will cause errors if not properly formatted. For example, while testing I was expecting an NVP entry not found exception for all Item, Mod, and Unit, but Unit gave a different error it gave "rapidjson internal assertion failure: IsObject()." This isn't exactly what I was expecting. I found that I had left an extra ',' at then of the last JSON entry (the last entry should have no trailing comma). During development, I also got some formatting errors that were giving me grief.