Serialize (#13)
*******New Keyword Constructor functionality via JSON Serialization********
This update includes new functionality that will allow me to create new GameObject and child classes using Keyword Constructors. 
* Item now has a Item(string keyword) constructor.
* Unit now has a Unit(string keyword) constructor.
* Mod now has a Mod(string keyword, GameObject owner) constructor.
*GameObject doesn't currently have one because it doesn't really need one yet.
* These keyword constructs use a custom scheme where a Parent object only stores a keyword for any of its children. So a json entry wouldn't have a complete child object, just the keyword. In the Parent constructor, it will deserialize the keywords, then use those to call the appropriate child object keyword constructor. I did this to avoid nesting too many layers of objects in any of the given json files. This way, the json files only store one kind of object. If you want to add a child object to an entry, you would just add a keyword to the appropriate attribute, then create that new child object in its respective json file using the same key.
*Cereal v1.3.2 has been implemented into the project. Currently, GameObjects are the only objects using it for JSON serialization. Eventually, this functionality will be expanded for saving the game.
* Unit.json, Mod.json, Item.json, and GameObject.json are new files that live in resources/jsonArchives. If you want to create a new object, do it here.
*****************************************************************************

* Working copy of serialize branch. v0.01

This works, but it really needs a review. I had several memory issues while implementing the serialize functions and their new constructors. I'll write more about that later. For now, this works and I want to publish it so I have something to fall back to.

* Serializer v0.1

Cleaned things up around the serializer functionality and just general cleanup of some of the GameObject child classes. Restructured them a bit, added better notes about the functions, and what not. Now, the game object classes (minus Player) all have a new constructor that takes a keyword, uses that keyword to find a json entry in a json, then makes a new object based on that found entry. Another feature to make it easier to create objects as json entries, is that when referencing another object, say a Mod child on an Item, you only need to reference it's keyword. It will then use the keyword constructor to make that child object.

What this does not have is any sort of exception handling, which is going to be required if most of the objects are going to be made in json entries so they can only be validated during runtime when the file is accessed.

* serealizer v1.0

Now that I've implemented the try/catch exceptions, I'm OK with calling this feature complete for now. I may eventually change the catch logic for these so they don't completely terminated the program. Maybe they could possible do an auto save or something? I'm not sure. But for now, it will throw an exception and terminate the program with a exit code of 1.
There is one thing to be mindful of when messing with any of the serialization stuff, and that's JSON files will cause errors if not properly formatted. For example, while testing I was expecting an NVP entry not found exception for all Item, Mod, and Unit, but Unit gave a different error it gave "rapidjson internal assertion failure: IsObject()." This isn't exactly what I was expecting. I found that I had left an extra ',' at then of the last JSON entry (the last entry should have no trailing comma). During development, I also got some formatting errors that were giving me grief.

* removing tests