Changing Item.modList from storing Mod objects to storing Mod object smart pointers.
Made changes to Item::modList, to now be a vector of Mod shared_ptrs instead of storing Mods directly. Had to update the addMods/removeMods methods in the Unit class because it was expecting a list of Mod objects instead of a list of Mod shared_ptrs.
Also found a bug in updateMods that was zeroing out the modded stat values.

This change brings Item up to the standard that all other GameObjects (as child or reference) are stored as shared_ptrs.