Rule and RuleSet classes implemented v.1
Slide deck has gone really well and has helped me plan out a lot of my next few steps. The first thing I decided to work on was the Rule and RuleSet classes. I wasn't exactly sure how to structure these, so maybe I did it wrong, but they are each their own class because I wanted to be able to do check functions for each of them.
What are they and how do they work? Let's start with a Rule. A Rule is mostly three components: a targetTag, a comparison operator, and a count. The GameMode will call tagCheck() which take the targetTag from the Rule, and calls tagCount() on the GameObject. It will then evaluate that result using the comparison operator and the rule's count. It will then return T/F based on the result. For example, say you have a Sword with one "Weapon" tag and a Rule "Weapon==0". GameMode calls GameMode.tagCheck(Sword, Weapon==0) which will return true if the count of "Weapon" tags on the GameObject is equal to 0. It is not, so it returns false.
A RuleSet is a way to do multiple tag checks at once and evaluate the results of those tag checks with a logical operator of either AND or OR. To use it, the GameMode calls multiTagCheck() which takes the GameObject and a RuleSet, and calls tagCheck on each of the Rules. Let's say there is a RuleSet with two Rules; "Weapon==0", "Weapon<2" with a logical operator of OR. If you called this on the same Sword, the "Weapon==0" tagCheck would return false, but the "Weapon<2" tagCheck would return true. Because the logical operator for the RuleSet is OR and "Weapon<2" is true, the RuleSet would return true.
I had initially thought to add the tagCheck and multiTagCheck to the GameObject, but I've decided against that for now. That decision is up for change, but this will work for the time being.