Consuambles v.1 (#18)
Consumables were relatively easy to implmement, though I did spend a couple hours looking for a really basic bug that I made. (I thought it was really complex so I looked past it for the longest time. Turned out to be a simple issue of accessing null memory.) Anyway, consumables are primarily a way to pass around buffs without attaching them to an item, giving much greater flexibility for adding mods, tags, and eventually abilities.
We are also going to be using Consumables to level characters, but there are going to need to be several updates before then. Two of those updates are going to be major refactor of the code where I'll move stats and items into maps.