v01 - statMap
I've created the statMap and the getStat() function. I was able to get things working properly. One thing of note, is that when getSat cannot find a stat it is supposed to return NULL, but it returns 0 instead. I suppose that is the same numerical value, but it isn't really want I wanted. I'm not sure if I should leave it like this, or just have it return 0 instead. I'm going to leave this for now and revisit it later.

In doing the statMap update, I quickly came into a big issue. Mods use ENUMs so should I have statMap use ENUMS or strings? Mods using ENUMs kind of sucks TBH because its ENUMs are numerical. To make statMap ENUMs, I would need to move the statType enum out of the mod class and into the GameObject class (so that both can have access to them easily). Or I can use STRINGS, which are more human readable in both the code and the JSON. I decided to go with strings.
So I made alterations to the MOD code and JSON files, I commented out how items show their stats (several classes now need some UI attention), and I updated the updateMod function in the Unit class. The updateMod function, was actually the only thing that I needed to do and it really wasn't that bad.
One other thing, I'm actually going to turn off MULTIPLY type mods for right now. I wasn't really sold on how I did this math before and doing this operation is going to be more difficult now as we move to dynamic stats. So for now, I'm going to comment this out while I continue updating Mod functionality.
At this point, I the statMap working and updating properly, like it did before. I've started making test functions in main. I actually did thsi Test Driven Development style, where I wrote out the test function in main first (just a bunch of simple operations to test each part and make sure it was all working), and then coded each part to match the test. It was good and I think I'll do that for the weapons as well. I should have been doing this all along as I was adding more and more functionality. I should probably move the test functions to their own file or something but we'll get there when we have to.