Initial Implementation complete. (#20)
Implementing itemMap was very similar process to implementing the statMap. I wrote up a test, code it to match, and was able to catch some bugs. So that went well and was relatively quick to do. Unlike statMap, however, GM did requre an update, specifically in the equip unit loop. Other than that, it was straight forward. I had to update the jsons, and the add/remove item functions. Anyway, that's it for now! This has been good work.
Next step is to update this merge this into QA and then clean up the code. It's messy. I should expand the tests, and then do the cleanup, just so that I know if there are any bugs before i start. Or to make sure that I don't make any bugs while cleaning up. lol