Some Refactoring and Clean Up
After recently reading through Andrey Karpov's "60 terrible tips for a C++ developer", ( https://pvs-studio.com/en/blog/posts/cpp/1053/ ) I decided that it was time to review my code with a critical eye, and do some refactoring and clean up. A lot of the changes are to the formatting and style of the classes, intended to help with readability and clarity. Such as "tabelizing" my code. You'll see this in a lot of the constructors. A major theme through Karpov's article was about forgoing concise code for readable code. Tabelizing is going to increase the overall length of my code, but it's uniformity should make it easier to spot errors.
Another thing that I ended up doing was removing all of the ENUMs. This was a tip from Karpov's article, but I also found the ENUMs lacking as they showed up as ints in my JSON instead of their more meaningful string.
Units and Item both required significant work, particularly in removing the use of ENUMs. I also went through them and cleaned up a bunch of the declared stats that are now obselete thanks to statMaps. This last part was actually a cleanup from the implementation of statmaps. Oh, and I also updated the stat choice map for items, so that they give you some additional details about the item.

At this point, I'm relatively happy with the changes that I've made. However, most of the work that was done was only done for the GameObject classes. The other important classes of PrntScrn and Inventory did not get any love. I'm actually a little reticent to open them back up as I am not sure I'm going to really like what I see. lol. Any changes in there might spawn more changes and I'm not sure any of that stuff is high priority enough to consider right now. I'll probably do that when I get closer to the campaign loop being completed and I want to make sure that the UI looks good.
The next step that I'm going to do is a big one for me, personally. I'm going to move my project to C++ 20. This is another tip from Karpov, who encourages you to use the latest version of C++ because it has the most modern features and runs better/more efficient than older version. Most people, including me, have been told that you should use the older versions because your program is more readily compileable elsewhere. He said that this is not something that most projects need to consider. And if you do need to consider it, you are probably aware of that already. So, based on that advice, I'm moving to C++20 so I can utilize some more modern features. I'm not entirely sure what all I'll get (I'm currently use 14), so there may be more refactors in the future if I find anything cool.