Events v1.0
Today I'm going to design my Events, a publisher class, Ability Class and Event Validation. The first two will be fundamental for all events, and the second two are going to be needed for non-combat triggers and can be used in the campaign loop, so they are the priority to get Decrement working how I want it.

To begin, I copied over most of the code by Shmuel Zang for his event system using standard C++. I made some alterations, but kept most of it. (I have extensive notes on how it works and a link to the original article in my project folder on james' google drive.) Next, I need to design my publisher class, which will hold the events for objects to subscribe to.
In thinking about the publisher class, I'm thinking that instead, I actually make this part of the GameState class.

So I created a test in the consoleApp1 file and it worked really well. I think the only thing I'm kind of concerned about at this point is going to be setting these up correctly with the right types. There will probably be some struggles with that, but that's OK. I'll learn it with experience.