IAbility Interface
This is working code that I have so far for the IAbility functionality. This version includes the basics but it's not fully functional. I still need to add it to the Buff and Unit classes. However, this several files will be the basis for the IAbility class.

sys/IAbilityDirector is the director of the IAbilityFactory Interface. This class is used by the client and has a switch service called "getIAbility" which will take a string as an argument, then use that to find the proper IAbilityFactory subclass to call. Which will then return an IAbility product to the client.

interfaceClasses/IAbilityFactory.h is an abstract class that has at least one method "makeIAbility", which will have to be implemented by any class that extends the interface.

interfaceClasses/IAbilityClasses/[IAbilityName].h->[IAbilityFactory] - A concrete factory (a child class of Base_Ability that implements the IAbilityFactory interface class) that holds the custom code for the [IAbilityName] class. This factory class has to implement the "makeIAbility" method that takes a GameObject pointer to an owner, uses that to create the [IAbilityName] subclass object, then returns the object cast as an IAbility product.

interfaceClasses/IAbilityClasses/[IAbilityName].h->[IAbilityName] a concrete product (a child class of Base_Ability that implements the IAbility interface and is produced by the corresponding [IAbilityName]Factory) that has the custom event functionality and implements the "update" method from the IAbility interface.

interfaceClasses/IAbility.h is an abstract interface class. Classes using this interface will have to implement an "update" method that takes a pointer to the current GameState, and registers the IAbility with it's corresponding Event held on the GameState's list of Events.

gameObjectClasses/Base_Ability.h: is parent class for the bare minimum amount of data needed for an ability. It's also how to save and load an ability from the Ability.json